+++ cromwell.example.backends/cromwell.examples.conf
@@ -157,15 +157,21 @@ call-caching {
   #   # The call caching blacklist cache is off by default. This cache is used to blacklist cache hit paths based on the
   #   # prefixes of cache hit paths that Cromwell has previously failed to copy for permissions reasons.
   #   enabled: true
-  #   # A blacklist grouping can be specified in workflow options which will inform the blacklister which workflows
-  #   # should share a blacklist cache.
-  #   grouping-workflow-option: call-cache-blackist-group
   #   # Guava cache concurrency.
   #   concurrency: 10000
   #   # How long entries in the cache should live from the time of their last access.
   #   ttl: 20 minutes
   #   # Maximum number of entries in the cache.
   #   size: 1000
+  #
+  #   # A blacklist grouping can be specified in workflow options which will inform the blacklister which workflows
+  #   # should share a blacklist cache.
+  #   groupings {
+  #     workflow-option: call-cache-blackist-group
+  #     concurrency: 10000
+  #     ttl: 2 hours
+  #     size: 1000
+  #   }
   # }
 }
